Integrating product base price for remove and create price commands. Fixing bugs related to product controller.

+++ src/products/application/commands/remove-price.command.ts
@@ -3,6 +3,8 @@ import { Inject, Injectable } from '@nestjs/common';
 import { ProductRepository } from '@/src/products/core/interfaces/ProductRepository.repository';
 import { ProductAggregate } from '@/src/products/core/aggregates/product.aggregate';
 import { ProductEntity } from '@/src/products/core/entities/product.entity';
+import { ProductPriceObjectValue } from '../../core/value-objects/product-price.object-value';
+import { BusinessRuleException } from '@/src/shared/errors/BusinessRuleException';
 
 @Injectable()
 export class RemovePriceCommand {
@@ -17,14 +19,20 @@ export class RemovePriceCommand {
       throw new Error(`Product with id ${id_product} does not exist.`);
     }
 
-    const aggregate = new ProductAggregate(products[0]);
+    const productAggregate = new ProductAggregate(products[0]);
+
+    const priceProductBase:ProductPriceObjectValue = productAggregate.getProductBasePrice();
+
+    if(id_product_price === priceProductBase.id_product_price) {
+      throw new BusinessRuleException('Base product price cannot be removed.')
+    }
 
     /*
       Business rules enforced inside aggregate:
       - Product must be active
       - Price must exist on this product
     */
-    aggregate.removePrice(id_product_price);
+    productAggregate.removePrice(id_product_price);
 
     await this.productRepository.deleteProductPrice(id_product_price);
   }
